Define payload input as a standalone component
