+++ website/frontend/src/components/blocks/FormPayload/FormPayload.module.scss
@@ -1,0 +1,3 @@
+.switch {
+  margin-bottom: 24px;
+}
