+++ website/frontend/src/components/pages/State/State.module.scss
@@ -142,7 +142,3 @@
     box-shadow: 0px 6px 10px rgba(0, 0, 0, 0.14), 0px 1px 18px rgba(0, 0, 0, 0.12), 0px 3px 5px rgba(0, 0, 0, 0.2);
   }
 }
-
-.switch {
-  margin-bottom: 24px;
-}
